# backend/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Astro Guide API",
    "description": "API for astrological and Human Design calculations and analysis",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    }
  ],
  "paths": {
    "/api/test": {
      "get": {
        "summary": "Test server status",
        "responses": {
          "200": {
            "description": "Server is running",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Server is running!"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/user": {
      "post": {
        "summary": "Create new user and generate analysis",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "birthday", "birthplace", "jobTitle"],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "John Doe"
                  },
                  "birthday": {
                    "type": "string",
                    "format": "date",
                    "example": "1990-01-01"
                  },
                  "birthtime": {
                    "type": "string",
                    "example": "12:00"
                  },
                  "birthplace": {
                    "type": "string",
                    "example": "New York, USA"
                  },
                  "jobTitle": {
                    "type": "string",
                    "example": "Software Engineer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "userId": {
                      "type": "integer",
                      "example": 1
                    },
                    "name": {
                      "type": "string",
                      "example": "John Doe"
                    },
                    "astroData": {
                      "type": "object"
                    },
                    "hdData": {
                      "type": "object"
                    },
                    "resonance": {
                      "type": "string"
                    },
                    "archetype": {
                      "type": "string"
                    },
                    "chartUrl": {
                      "type": "string"
                    },
                    "chatResponse": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "All fields are required"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Server error: Failed to process request"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/chat": {
      "post": {
        "summary": "Send message and get personalized response",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["userId", "message"],
                "properties": {
                  "userId": {
                    "type": "integer",
                    "example": 1
                  },
                  "message": {
                    "type": "string",
                    "example": "What career path suits me best?"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chat response generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "response": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "User ID and message are required"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "User not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Server error: Failed to process chat"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
} 